error-handler/src/main/java/uz/cherevichenko/exception_handler/global_exception_handler/GlobalExceptionHandler.java
package uz.cherevichenko.exception_handler.global_exception_handler;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.context.request.WebRequest;

import uz.cherevichenko.exception_handler.config.ErrorHandlerProperties;
import uz.cherevichenko.exception_handler.exception.AccessDeniedException;
import uz.cherevichenko.exception_handler.exception.InvalidInputException;
import uz.cherevichenko.exception_handler.exception.ResourceNotFoundException;
import uz.cherevichenko.exception_handler.model.ErrorDetails;

@ControllerAdvice
public class GlobalExceptionHandler {

    @Autowired
    private ErrorHandlerProperties errorHandlerProperties;

    @ExceptionHandler(Exception.class)
    public ResponseEntity<?> handleGlobalException(Exception ex, WebRequest request) {
        if (errorHandlerProperties.isEnabled()) {
            ErrorDetails errorDetails = new ErrorDetails(HttpStatus.INTERNAL_SERVER_ERROR.value(), ex.getMessage(), request.getDescription(false));
            return new ResponseEntity<>(errorDetails, HttpStatus.INTERNAL_SERVER_ERROR);
        }
        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
    }

    @ExceptionHandler(ResourceNotFoundException.class)
    public ResponseEntity<?> handleResourceNotFoundException(ResourceNotFoundException ex, WebRequest request) {
        if (errorHandlerProperties.isResourceNotFoundEnabled()) {
            ErrorDetails errorDetails = new ErrorDetails(HttpStatus.NOT_FOUND.value(), ex.getMessage(), request.getDescription(false));
            return new ResponseEntity<>(errorDetails, HttpStatus.NOT_FOUND);
        }
        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
    }

    @ExceptionHandler(InvalidInputException.class)
    public ResponseEntity<?> handleInvalidInputException(InvalidInputException ex, WebRequest request) {
        if (errorHandlerProperties.isInvalidInputEnabled()) {
            ErrorDetails errorDetails = new ErrorDetails(HttpStatus.BAD_REQUEST.value(), ex.getMessage(), request.getDescription(false));
            return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);
        }
        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
    }

    @ExceptionHandler(AccessDeniedException.class)
    public ResponseEntity<?> handleAccessDeniedException(AccessDeniedException ex, WebRequest request) {
        if (errorHandlerProperties.isAccessDeniedEnabled()) {
            ErrorDetails errorDetails = new ErrorDetails(HttpStatus.FORBIDDEN.value(), ex.getMessage(), request.getDescription(false));
            return new ResponseEntity<>(errorDetails, HttpStatus.FORBIDDEN);
        }
        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
    }
}
